gh-421 - split up gaffer.py into smaller components
